+++ pages/api/processos/[cliente_id].ts
@@ -7,7 +7,7 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
             const cliente_id = req.query.cliente_id;
 
             if (!cliente_id || isNaN(Number(cliente_id))) {
-                return res.status(400).json({ error: "Necessário informar um id de cliente válido." });
+                throw new Error("Necessário informar um id de cliente válido.");
             }
 
             const sql = `
@@ -26,12 +26,8 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
                 WHERE c.id = ?
                 ORDER BY c.nome ASC;
             `;
-            
-            const rs_processos = await query(sql, [cliente_id]);
 
-            if (rs_processos.length === 0) {
-                return res.status(404).json({ error: "Nenhum processo encontrado para o cliente informado." });
-            }
+            const rs_processos = await query(sql, [cliente_id]);
 
             res.status(200).json(rs_processos);
         } else {
@@ -39,6 +35,6 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
             res.status(405).end(`Method ${req.method} Not Allowed`);
         }
     } catch (erro) {
-        res.status(500).json({ error: erro.toString() });
+        res.status(400).json({ error: erro.toString() });
     }
 }
